small fixes to release scripts

+++ scripts/release.sh
@@ -27,19 +27,25 @@ if [ $CURL_STATUS -ne 0 ]; then
 fi
 if [ $STATUS -ne 201 ]; then
   echo "Could not create release; GitHub returned $STATUS"
+  cat release.json
   rm -f release.json
   exit 1
 fi
 
-id=`cat release.json | python -c 'import json, sys; print json.load(sys.stdin)["id"]'`
-rm -f release.json
-echo "Release $id created"
+id=`cat release.json | python -c 'import json, sys; print(json.load(sys.stdin)["id"])'`
+if [ ! -z $id ]; then
+  echo "Release $id created"
+  rm -f release.json
+else
+  echo "Release not found; see release.json"
+  exit 1
+fi
 
 for file in "$@"; do
   name=`basename $file`
   echo "Uploading release asset $name"
 
-  STATUS=$(curl https://api.github.com/uploads/repos/megalord/saml/releases/$id/assets?name=$name -H "Authorization: token $GITHUB_TOKEN" -H 'Accept: application/vnd.github.v3+json' -H 'Content-Type: application/gzip' --data-binary "@$file" -s -o /dev/null -w "%{http_code}")
+  STATUS=$(curl https://uploads.github.com/repos/megalord/saml/releases/$id/assets?name=$name -H "Authorization: token $GITHUB_TOKEN" -H 'Accept: application/vnd.github.v3+json' -H 'Content-Type: application/gzip' --data-binary "@$file" -s -o /dev/null -w "%{http_code}")
   CURL_STATUS=$?
   if [ $CURL_STATUS -ne 0 ]; then
     echo "Could not create release asset $name; curl failed with error code $CURL_STATUS"
